requirements.txt
asgiref==3.5.1
Babel==2.10.1
Django==3.2.13
django-extra-views==0.13.0
django-haystack==3.2.1
django-oscar==3.1
django-oscar-api==3.1.1
django-phone-field==1.8.1
django-phonenumber-field==5.2.0
django-tables2==2.3.4
django-treebeard==4.4
django-widget-tweaks==1.4.12
djangorestframework==3.13.1
factory-boy==3.2.1
Faker==13.11.1
flake8==4.0.1
mccabe==0.6.1
phonenumbers==8.12.48
Pillow==9.1.0
purl==1.6
pycodestyle==2.8.0
pycountry==22.3.5
pyflakes==2.4.0
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
six==1.16.0
sorl-thumbnail==12.6.3
sqlparse==0.4.2
whitenoise==6.2.0
